+++ core-ng/src/main/java/core/framework/internal/web/websocket/WebSocketContextImpl.java
@@ -8,6 +8,7 @@ import org.slf4j.LoggerFactory;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
 /**
@@ -15,8 +16,9 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public class WebSocketContextImpl implements WebSocketContext {
     private final Logger logger = LoggerFactory.getLogger(WebSocketContextImpl.class);
-    private final Map<String, Map<String, Channel<?>>> rooms = new ConcurrentHashMap<>();
     private final Map<String, Channel<?>> channels = new ConcurrentHashMap<>();
+    private final Map<String, Map<String, Channel<?>>> rooms = new ConcurrentHashMap<>();
+    private final Map<String, Channel<?>> keys = new ConcurrentHashMap<>();
 
     @SuppressWarnings("unchecked")
     @Override
@@ -40,6 +42,13 @@ public class WebSocketContextImpl implements WebSocketContext {
         return results;
     }
 
+    @Override
+    public <V> Optional<Channel<V>> key(String key) {
+        @SuppressWarnings("unchecked")
+        Channel<V> channel = (Channel<V>) keys.get(key);
+        return Optional.ofNullable(channel);
+    }
+
     void join(ChannelImpl<?, ?> channel, String room) {
         logger.debug("join room, channel={}, room={}", channel.id, room);
         channel.rooms.add(room);
@@ -60,7 +69,14 @@ public class WebSocketContextImpl implements WebSocketContext {
     void remove(ChannelImpl<?, ?> channel) {
         channels.remove(channel.id);
         for (String room : channel.rooms) {
-            rooms.get(room).remove(channel.id);
+            Map<String, Channel<?>> roomChannels = rooms.get(room);
+            roomChannels.remove(channel.id);
+            if (roomChannels.isEmpty()) rooms.remove(room); // cleanup room if has no channels
         }
+        if (channel.key != null) keys.remove(channel.key);
+    }
+
+    void updateKey(ChannelImpl<?, ?> channel) {
+        keys.put(channel.key, channel);
     }
 }