+++ core-ng/src/main/java/core/framework/web/websocket/WebSocketContext.java
@@ -1,6 +1,7 @@
 package core.framework.web.websocket;
 
 import java.util.List;
+import java.util.Optional;
 
 /**
  * @author neo
@@ -9,4 +10,6 @@ public interface WebSocketContext {
     <V> List<Channel<V>> all();
 
     <V> List<Channel<V>> room(String name);
+
+    <V> Optional<Channel<V>> key(String key);
 }